fix(gossipsub): don't panic on mcache while adding fanout peers
Fanout peers may not be been added correctly because the scoring function was not negated. This PR changes it to be more consistent with other parts of the code base.

Also, users can set the memcache size to be 0. When this happens, a few panics occur. I've added statements to prevent this.

Pull-Request: #6095.