fix(gossipsub): Always construct mesh
There is a case where we subscribe to a topic but never create a mesh. This can happen for example when we set the mesh degree to 0, then we don't add any peers and the mesh never gets constructed. We then never send subscriptions to peers and essentially the subscription never existed.

This PR ensures the mesh is always constructed when the user subscribes.

This PR also improves the logging to downgrade some logs and remove unnecessary logs.

Pull-Request: #6100.