fix(gossipsub): Remove penalty for duplicate publish messages
We had a mechanism designed to punish peers that send messages to us with the same message-id. This was a mitigation to prevent peers from sending us back messages we just published. However, in content-based networks, there is often the scenario (which is not explicitly forbidden in gossipsub) where there can be multiple publishers of the same message.

This mechanism was downscoring honest nodes in this system. This PR removes this mechanism. If we want to have something like this, we can tie it now to nodes sending messages after recieving IDONTWANT's from us. This has its own pitfalls and I leave for a future PR.

cc @dknopik

Pull-Request: #6112.